% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronbach_alpha.R
\name{cronbach_alpha}
\alias{cronbach_alpha}
\title{Compute Cronbach alpha with confidence interval}
\usage{
cronbach_alpha(Data, ci = TRUE, ci_lvl = 0.95)
}
\arguments{
\item{Data}{\emph{data.frame} or \emph{matrix}, item data, \code{NA} gets excluded
automatically.}

\item{ci}{\emph{logical}, whether to compute CI or not. Defaults to \code{TRUE}.}

\item{ci_lvl}{\emph{numeric} ranging from 0 to 1, a confidence level to construct
CI for. Defaults to \code{.95}.}
}
\value{
A list with \eqn{\alpha} estimate and optionally CI.
}
\description{
So far internal function.
}
\examples{
ShinyItemAnalysis:::cronbach_alpha(HCI[, 1:20])

}
\keyword{internal}
