% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMedical.R
\docType{data}
\name{dataMedical}
\alias{dataMedical}
\title{Dichotomous Data Set of Admission Test to Medical School}
\format{A \code{dataMedical} data frame consists of 3,204 observations on
100 variables. The columns represent dichotomously scored items of the test.}
\usage{
data(dataMedical)
}
\description{
The \code{dataMedical} data set consists of the responses of 3,204 subjects
to admission test to medical school in the Czech republic.
It contains 100 items A correct answer is coded as 1 and incorrect answer as 0.
}
\author{
Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr

Jakub Houdek \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
houdek.james@gmail.com \cr
}
\seealso{
\code{\link{dataMedicaltest}}, \code{\link{dataMedicalkey}}
}
\keyword{datasets}

