% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMedicalkey.R
\docType{data}
\name{dataMedicalkey}
\alias{dataMedicalkey}
\title{Key of Correct Answers for \code{dataMedicaltest} Data Set}
\format{A data frame with 100 values representing correct answers to items of \code{dataMedicaltest}
data set. For more details see \code{\link{dataMedicaltest}}.}
\usage{
data(dataMedicalkey)
}
\description{
The \code{dataMedicalkey} is a vector of factors representing correct answers of
\code{dataMedicaltest} data set.
}
\author{
Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr

Jakub Houdek \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
houdek.james@gmail.com \cr#'
}
\seealso{
\code{\link{dataMedical}}, \code{\link{dataMedicaltest}}
}
\keyword{datasets}

