% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMedicaltest.R
\docType{data}
\name{dataMedicaltest}
\alias{dataMedicaltest}
\title{Data Set of Admission Test to Medical School}
\format{A \code{dataMedicaltest} data frame consists of 3,204 observations on 100 variables. The columns represent answers of subject to items of an admission test.}
\usage{
data(dataMedicaltest)
}
\description{
The \code{dataMedicaltest} data set consists of the responses of 3,204 subjects to multiple-choice admission test to medical school in Czech republic. It contains 100 items . Possible answers were A, B, C, D, while any combination of these can be correct.
}
\author{
Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr

Jakub Houdek \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
houdek.james@gmail.com \cr#'
}
\seealso{
\code{\link{dataMedical}}, \code{\link{dataMedicalkey}}
}
\keyword{datasets}

