% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_method.R
\name{delta_ses}
\alias{delta_ses}
\title{Get standard errors using delta method approximation}
\usage{
delta_ses(formula, mean, cov)
}
\arguments{
\item{formula}{a list of RHS-only formula(s)}

\item{mean}{estimated mean of X}

\item{cov}{estimated variance-covariance matrix of X}
}
\value{
SEs vector
}
\description{
Get standard errors using delta method approximation
}
\keyword{internal}
