% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCIdata.R
\docType{data}
\name{HCIdata}
\alias{HCIdata}
\title{Homeostasis Concept Inventory Dataset}
\format{
\code{HCIdata} is a \code{data.frame} consisting of 669 observations on the 47 variables.
\describe{
\item{A1-A20}{multiple-choice items of the HCI test}
\item{QR1-QR20}{scored items of the HCI test, \code{"0"} incorrect, \code{"1"} correct}
\item{total}{total score}
\item{gender}{gender membership, \code{"M"} males, \code{"F"} females, \code{"none"} undisclosed}
\item{major}{identifier whether students planning to major in the life sciences}
\item{yearc5}{study year}
\item{minority}{minority membership, \code{"maj"} majority, \code{"min"} minority, \code{"none"} undisclosed}
\item{EnglishF}{identifier whether English is the student's first language}
\item{typeS}{course type, \code{"allied"} allied health, \code{"majors"}  physiology courses for science majors,
\code{"mixed majors"} courses for nonmajors}
\item{typeSCH}{type of school, \code{"AC"} associate's college, \code{"BCAS"}  baccalaureate college: arts and
sciences focus, \code{"R1"} research university, \code{"MCU"} master's college and university}
}
}
\usage{
data(HCIdata)
}
\description{
\code{HCIdata} dataset consists of the responses of 669 students (405 males,
246 females, 18 without gender specification) to Homeostasis Concept Inventory (HCI)
multiple-choice test. It contains answers to 20 multiple-choice items, scored items,
total score, gender membership, identifier whether students plan to major in science, study year,
minority membership, identifier whether English is the student's first language, and type of school.
}
\references{
McFarland, J. L., Price, R. M., Wenderoth, M. P., Martinkova, P., Cliff, W., Michael, J., ... & Wright, A. (2017).
Development and validation of the homeostasis concept inventory. CBE-Life Sciences Education, 16(2), ar35.
}
\seealso{
\code{\link{HCI}}, \code{\link{HCItest}}, \code{\link{HCIkey}}
}
\author{
Jenny L. McFarland \cr
Biology Department, Edmonds Community College
}
\keyword{datasets}
