% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_HCIgrads.R
\docType{data}
\name{HCIgrads}
\alias{HCIgrads}
\title{Homeostasis concept inventory dataset of graduate students}
\format{
\code{HCIgrads} is a \code{data.frame} consisting of 10 observations
on the 42 variables.
\describe{
\item{A1-A20}{Multiple-choice items of the HCI test. }
\item{QR1-QR20}{Scored items of the HCI test, \code{"0"} incorrect,
\code{"1"} correct. }
\item{total}{Total test score. }
}
}
\usage{
HCIgrads
}
\description{
\code{HCIgrads} dataset consists of the responses of 10 graduate
students to Homeostasis Concept Inventory (HCI) multiple-choice test. It
contains answers to 20 multiple-choice items, scored items, and total test
score.
}
\references{
McFarland, J. L., Price, R. M., Wenderoth, M. P., Martinkova, P.,
Cliff, W., Michael, J., ... & Wright, A. (2017). Development and validation
of the homeostasis concept inventory. CBE-Life Sciences Education, 16(2),
ar35. \doi{10.1187/cbe.16-10-0305}
}
\seealso{
\link{HCI} for HCI dichotomous dataset\cr
\link{HCItest} for HCI multiple-choice dataset\cr
\link{HCIkey} for key of correct answers for HCI\cr
\link{HCIdata} for HCI full dataset\cr
\link{HCIlong} for HCI in a long format\cr
\link{HCIprepost} for HCI pretest and posttest scores\cr
\link{HCItestretest} for HCI test-retest dataset\cr
}
\author{
Jenny L. McFarland \cr Biology Department, Edmonds Community College
}
\keyword{datasets}
