% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCItestretest.R
\docType{data}
\name{HCItestretest}
\alias{HCItestretest}
\title{Homeostasis Concept Inventory Dataset}
\format{
\code{HCItestretest} is a \code{data.frame} consisting of 90 observations on the 44 variables.
\describe{
\item{A1-A20}{multiple-choice items of the HCI test}
\item{QR1-QR20}{scored items of the HCI test, \code{"0"} incorrect, \code{"1"} correct}
\item{test}{identifier of test vs retest, \code{"test"} test, \code{"retest"} retest after }
\item{total}{total score}
\item{gender}{gender membership, \code{"M"} male, \code{"F"} female.}
\item{major}{identifier whether student plans to major in the life sciences.}
}
}
\usage{
data(HCItestretest)
}
\description{
\code{HCItestretest}) dataset consists of the responses of 45 students
to Homeostasis Concept Inventory (HCI). It contains answers to 20 multiple-choice items,
scored items, identifier of test/retest, total score, gender membership and identifier
whether students plan to major in life sciences.
The data are organized so that each pair of subsequent rows belongs to one student.
Students took no courses on homeostasis between the test and retest.
}
\references{
McFarland, J. L., Price, R. M., Wenderoth, M. P., Martinkova, P., Cliff, W., Michael, J., ... & Wright, A. (2017).
Development and validation of the homeostasis concept inventory. CBE-Life Sciences Education, 16(2), ar35.
}
\seealso{
\code{\link{HCItest}}, \code{\link{HCIkey}}
}
\author{
Jenny L. McFarland \cr
Biology Department, Edmonds Community College
}
\keyword{datasets}
