% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_HeightInventory.R
\docType{data}
\encoding{UTF-8}
\name{HeightInventory}
\alias{HeightInventory}
\title{Height inventory dataset}
\format{
\code{HeightInventory} is a \code{data.frame} consisting of 4,885 observations
on the 28 variables. First 26 variables are responses on scale \code{"1"}
strongly disagree, \code{"2"} disagree, \code{"3"} agree, \code{"4"} strongly agree.
Items 14 - 26 were reverse-coded, so that all items are scored in the same
direction. Names of these items start with \code{"R-"}. Original item number
and English wording is provided below.

\describe{
\item{ShortTrousers}{1. A lot of trousers are too short for me.}
\item{TallerThanM}{2.	I am taller than men of my age.}
\item{TallerThanF}{3.	I am taller than women of my age.}
\item{HeightForBasketball}{4.	I have an appropriate height for playing basketball or volleyball.}
\item{AskMeToReach}{5.	Other people sometimes ask me to reach something for them.}
\item{CommentsTall}{6.	I am used to hearing comments about how tall I am.}
\item{ConcertObstructs}{7.	At concerts, my stature usually obstructs other people’s views.}
\item{ShortBed}{8.	Ordinary beds are too short for me.}
\item{TopShelfEasy}{9.	I can easily take wares from top shelves at a store.}
\item{CrowdViewComf}{10.	In a crowd of people, I still have a comfortable view.}
\item{ShortBlanket}{11.	Blankets and bedspreads rarely cover me completely.}
\item{BendToHug}{12.	When I want to hug someone, I usually need to bend over.}
\item{CarefulHead}{13.	I must often be careful to avoid bumping my head against a doorjamb or a low ceiling.}
\item{R-SmallerThanM}{14.	I am smaller than men of my age. (reversed)}
\item{R-StoolNeeded}{15.	I often need a stool to reach something other people could reach without one. (reversed)}
\item{R-PlayDwarf}{16.	I could play a dwarf. (reversed)}
\item{R-SmallerThanW}{17.	I am smaller than women of my age. (reversed)}
\item{R-NoticeSmall}{18.	One of the first things people notice about me is how small I am. (reversed)}
\item{R-OnTipToes}{19.	I often need to stand on the tip of my toes to get a better view. (reversed)}
\item{R-ClothChildSize}{20.	When I buy clothes, children’s sizes often fit me well. (reversed)}
\item{R-BusLegsEnoughSpace}{21.	I have enough room for my legs when traveling by bus. (reversed)}
\item{R-FasterWalk}{22.	I often need to walk faster than I’m used to in order to keep pace with taller people. (reversed)}
\item{R-AgeUnderestim}{23.	Because of my smaller stature, people underestimate my age. (reversed)}
\item{R-WishLowerChair}{24.	It would be more comfortable for me if chairs were made lower. (reversed)}
\item{R-UpwardLook}{25.	When talking to other adults, I have to look upwards if I want to meet their eyes. (reversed)}
\item{R-MirrorTooHigh}{26.	Some mirrors are placed so high up that I have to crane my neck to use them. (reversed)}
\item{gender}{Gender membership, \code{"M"} males, \code{"F"} females.}
\item{HeightCM}{Self-reported height in centimeters.}
\item{total}{Total score.}
}
}
\usage{
HeightInventory
}
\description{
\code{HeightInventory} dataset consists of the responses of 4,885 respondents
(1479 males, 3406 females) to a Height Inventory  (Rečka, 2018). It contains 26
ordinal items of self-perceived height rated on a scale \code{"1"} strongly disagree,
\code{"2"} disagree, \code{"3"} agree, \code{"4"} strongly agree, vector of self-reported
heights (in centimeters), and vector of gender membership. Total score is included
as the last variable, total score is NA for respondents who missed any item.
}
\note{
Thanks to Karel Rečka and Hynek Cígler for sharing this dataset.
}
\references{
Rečka, K. (2018). Height and Weight Inventory. Brno, Masaryk
University: Unpublished Master's thesis
}
\keyword{datasets}
