% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDIFirt.R
\name{plotDIFirt}
\alias{plotDIFirt}
\title{Function for characteristic curve of DIF IRT model}
\usage{
plotDIFirt(parameters, test = "Lord", item = 1)
}
\arguments{
\item{parameters}{numeric: data matrix or data frame. See \strong{Details}.}

\item{test}{character: type of statistic to be shown.}

\item{item}{numeric: number of item to be plotted.}
}
\description{
Plots characteristic curve of IRT model.
}
\details{
This function plots characteristic curve of DIF IRT model.
}
\examples{
\dontrun{


}


}
\author{
Adela Drabinova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
adela.drabinova@gmail.com \cr

Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr
}

