% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_empty_cols}
\alias{remove_empty_cols}
\title{Remove columns that are empty}
\usage{
remove_empty_cols(.data)
}
\arguments{
\item{.data}{data.frame}
}
\value{
cleaned df
}
\description{
Remove columns that are empty
}
\keyword{internal}
