\name{startShinyItemAnalysis}
\alias{startShinyItemAnalysis}
\title{
This function will start ShinyItemAnalysis application.
}
\description{
An interactive Shiny application for running Item Analysis.
}
\usage{
startShinyItemAnalysis()
}
\examples{
\dontrun{library(shiny)
library(shinyAce)
startShinyItemAnalysis()}
{
  }
}

