% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blis.R
\name{blis2blirt}
\alias{blis2blirt}
\title{Reparametrize fitted BLIS model to IRT}
\usage{
blis2blirt(fitted_model)
}
\arguments{
\item{fitted_model}{fitted BLIS model (object of class \linkS4class{BlisClass})}
}
\value{
IRT parametrization of BLIS model
}
\description{
Reparametrize fitted BLIS model to IRT
}
\keyword{internal}
