% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blis.R
\name{coef,BlisClass-method}
\alias{coef,BlisClass-method}
\title{Get Coefficients from a fitted BLIS model}
\usage{
\S4method{coef}{BlisClass}(
  object,
  ...,
  CI = 0.95,
  printSE = FALSE,
  IRTpars = FALSE,
  simplify = FALSE,
  labels = FALSE,
  mark_correct = labels
)
}
\arguments{
\item{object}{\emph{object of class \linkS4class{BlisClass}}, model fitted via \code{fit_blis}() or
\code{blis()}.}

\item{...}{Additional arguments. Not utilized at the moment.}

\item{CI}{\emph{numeric}, a width of the confidence intervals.}

\item{printSE}{\emph{logical}, print standard errors instead of CI? Defaults to
\code{FALSE}.}

\item{IRTpars}{\emph{logical}, convert slope intercept parameters into IRT
parameters (i.e. BLIRT)? Defaults to \code{FALSE}.}

\item{simplify}{\emph{logical}, return coefficients as a matrix, instead of list?
Defaults to \code{FALSE}. \emph{Not implemented yet.}}

\item{labels}{\emph{logical}, if \code{TRUE}, show response labels (e.g. "A", "B", "C")
instead of response numeric indices (e.g. 0, 1, 2). Defaults to \code{FALSE}.}

\item{mark_correct}{\emph{logical}, mark the correct response with an asterisk
symbol. Applicable only if \code{labels} is \code{TRUE} (in which case,
\code{mark_correct} defaults to \code{TRUE}).}
}
\value{
coefs
}
\description{
Get Coefficients from a fitted BLIS model
}
\examples{
"todo"
}
\seealso{
Other BLIS/BLIRT related: 
\code{\link{BlisClass-class}},
\code{\link{fit_blis}()},
\code{\link{get_orig_levels}()},
\code{\link{nominal_to_int}()},
\code{\link{obtain_nrm_def}()},
\code{\link{print.blis_coefs}()}
}
\concept{BLIS/BLIRT related}
