% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_CZmatura.R
\docType{data}
\name{CZmatura}
\alias{CZmatura}
\title{CZmatura dataset}
\format{
\code{CZmatura} is a \code{data.frame} consisting of 15,702 observations on
  75 variables.
  \describe{
  \item{SchType}{School type code. }
  \item{FirstAtt}{First attempt; \code{"1"} yes, \code{"0"} no. }
  \item{SchTypeGY}{School type gymnasium; \code{"1"} yes, \code{"0"} no. }
  \item{o1 -- o26.2}{Item answers. }
  \item{b1 -- b26}{Scored item answers. }
  \item{Total}{Total score, calculated as sum of item scores (0 - 50). }
  \item{IRTscore}{Score estimated from GPCM/2PL model. }
  \item{IRTscoreSE}{SE of score estimated from GPCM/2PL model. }
}
}
\usage{
data(CZmatura)
}
\description{
The \code{CZmatura} dataset comes from a matura exam in
  mathematics. The exam was assigned to students from Grade 13, at
  the end of their secondary education.
}
\seealso{
\code{\link{CZmaturaS}}
}
\keyword{datasets}
