% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMedical.R
\docType{data}
\name{dataMedical}
\alias{dataMedical}
\title{Dichotomous Data Set of Admission Test to Medical School}
\format{A \code{dataMedical} data frame consists of 3,204 observations on
100 variables. The columns represent dichotomously scored items of the test.}
\usage{
data(dataMedical)
}
\description{
The \code{dataMedical} data set consists of the responses of 3,204 subjects
to admission test to medical school in the Czech republic.
It contains 100 items. A correct answer is coded as 1 and incorrect answer as 0.
}
\author{
Cestmir Stuka \cr
First Faculty of Medicine, Charles University in Prague \cr

Martin Vejrazka \cr
First Faculty of Medicine, Charles University in Prague \cr

Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr

Jakub Houdek \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
The Faculty of Informatics and Statistics, University of Economics, Prague \cr
houdek.james@gmail.com \cr
}
\references{
Stuka, C. Vejrazka, M., Martinkova, P. Komenda, M. and Stepanek, L. (2016). The Use of  Test and Item Analisis for Improvement of Tests. Workshop held at conference MEFANET, 2016, Brno, Czech Republic.
}
\seealso{
\code{\link{dataMedicaltest}}, \code{\link{dataMedicalkey}}
}
\keyword{datasets}

