% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMedical.R
\docType{data}
\name{dataMedical}
\alias{dataMedical}
\title{Dichotomous Dataset of Admission Test to Medical School}
\format{A \code{dataMedical} is a \code{data.frame} consisting of 2,392 observations on
the following 102 variables. The first 100 columns represent dichotomously scored
items of the test. The 101st column is vector of gender membership; values 0 and 1
refer to males and females. The 102nd columns in criterion variable; value 1 means
that student studies standardly, 0 otherwise (e.g. leaving or interrupting studies).}
\usage{
data(dataMedical)
}
\description{
The \code{dataMedical} dataset consists of the responses of 2,392 subjects
(750 males, 1,633 females and 9 subjects without gender specification) to admission test to
a medical school. It contains 100 items. A correct answer is coded as 1 and incorrect
answer as 0. Missing answers were evaluated as incorrect, i.e. 0.
}
\references{
Stuka, C. Vejrazka, M., Martinkova, P. Komenda, M. & Stepanek, L. (2016). The Use of Test and Item Analisis for
Improvement of Tests. Workshop held at conference MEFANET, 2016, Brno, Czech Republic.
}
\seealso{
\code{\link{dataMedicaltest}}, \code{\link{dataMedicalkey}}, \code{\link{dataMedicalgraded}}
}
\author{
Cestmir Stuka \cr
First Faculty of Medicine, Charles University in Prague \cr

Martin Vejrazka \cr
First Faculty of Medicine, Charles University in Prague \cr

Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr

Jakub Houdek \cr
Institute of Computer Science, The Czech Academy of Sciences \cr

Adela Hladka  \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
hladka@cs.cas.cz \cr
}
\keyword{datasets}
