% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMedicalkey.R
\docType{data}
\name{dataMedicalkey}
\alias{dataMedicalkey}
\title{Key of Correct Answers for \code{dataMedicaltest} Dataset}
\format{A \code{data.frame} with 100 values representing correct answers to items of \code{dataMedicaltest}
dataset. For more details see \code{\link{dataMedicaltest}}.}
\usage{
data(dataMedicalkey)
}
\description{
The \code{dataMedicalkey} is a vector of factors representing correct answers of
\code{dataMedicaltest} data set.
}
\references{
Stuka, C. Vejrazka, M., Martinkova, P. Komenda, M. & Stepanek, L. (2016). The Use of  Test and Item Analisis for
Improvement of Tests. Workshop held at conference MEFANET, 2016, Brno, Czech Republic.
}
\seealso{
\code{\link{dataMedical}}, \code{\link{dataMedicaltest}}, \code{\link{dataMedicalgraded}}
}
\author{
Cestmir Stuka \cr
First Faculty of Medicine, Charles University in Prague \cr

Martin Vejrazka \cr
First Faculty of Medicine, Charles University in Prague \cr

Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr

Jakub Houdek \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
The Faculty of Informatics and Statistics, University of Economics, Prague \cr

Adela Hladka  \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
hladka@cs.cas.cz \cr
}
\keyword{datasets}
