% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMedicalkey.R
\docType{data}
\name{dataMedicalkey}
\alias{dataMedicalkey}
\title{Key of correct answers for \code{dataMedicaltest} dataset}
\format{
A vector with 100 values representing correct answers to items of
  \code{dataMedicaltest} dataset. For more details see
  \code{\link{dataMedicaltest}}.
}
\usage{
data(dataMedicalkey)
}
\description{
The \code{dataMedicalkey} is a vector of factors representing
  correct answers of \code{dataMedicaltest} dataset.
}
\references{
Stuka, C. Vejrazka, M., Martinkova, P. Komenda, M. & Stepanek, L.
(2016). The Use of  Test and Item Analisis for Improvement of Tests. Workshop
held at conference MEFANET, 2016, Brno, Czech Republic.
}
\seealso{
\code{\link{dataMedical}}, \code{\link{dataMedicaltest}},
  \code{\link{dataMedicalgraded}}
}
\author{
Cestmir Stuka \cr
First Faculty of Medicine, Charles University \cr

Martin Vejrazka \cr
First Faculty of Medicine, Charles University \cr

Patricia Martinkova \cr
Institute of Computer Science of the Czech Academy of Sciences \cr
\email{martinkova@cs.cas.cz} \cr
}
\keyword{datasets}
