% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCI.R
\docType{data}
\name{HCI}
\alias{HCI}
\title{Homeostasis Concept Inventory Dichotomous Dataset}
\format{\code{HCI} is a \code{data.frame} consisting of 651 observations on the 22 variables.
First 20 variables represent dichotomously scored responses to multiple-choice items (1 means
correct, 0 is incorrect). 21st column is a vector of gender membership; values 0 and 1
refer to males and females. 22nd column is a identificator whether students planning to
major in the life sciences.}
\usage{
data(HCI)
}
\description{
(\code{HCI}) dataset consists of the dichotomously scored responses of 651 students
(405 males, 246 females) to Homeostasis Concept Inventory multiple-choice test. It containts 20
items, vector of gender membership and identificator whether students plan to major.
}
\references{
McFarland, J. L., Price, R. M., Wenderoth, M. P., Martinkova, P., Cliff, W., Michael, J., ... & Wright, A. (2017).
Development and validation of the homeostasis concept inventory. CBE-Life Sciences Education, 16(2), ar35.
}
\seealso{
\code{\link{HCItest}}, \code{\link{HCIkey}}
}
\author{
Jenny L. McFarland \cr
Biology Department, Edmonds Community College
}
\keyword{datasets}
