% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCIkey.R
\docType{data}
\name{HCIkey}
\alias{HCIkey}
\title{Key of Correct Answers for Homeostasis Concept Inventory Dataset}
\format{A nominal vector with 20 values representing correct answers to items of \code{HCItest}
dataset. For more details see \code{\link{HCItest}}.}
\usage{
data(HCIkey)
}
\description{
The \code{HCIkey} is a vector of factors representing correct answers of
\code{HCItest} dataset.
}
\references{
McFarland, J. L., Price, R. M., Wenderoth, M. P., Martinkova, P., Cliff, W., Michael, J., ... & Wright, A. (2017).
Development and validation of the homeostasis concept inventory. CBE-Life Sciences Education, 16(2), ar35.
}
\seealso{
\code{\link{HCI}}, \code{\link{HCItest}}
}
\author{
Jenny L. McFarland \cr
Biology Department, Edmonds Community College
}
\keyword{datasets}
