% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_HCIkey.R
\docType{data}
\name{HCIkey}
\alias{HCIkey}
\title{Key of correct answers for homeostasis concept inventory dataset}
\format{
A nominal vector with 20 values representing correct answers to items
  of \code{HCItest} dataset. For more details see \code{\link{HCItest}}.
}
\usage{
data(HCIkey)
}
\description{
The \code{HCIkey} is a vector of factors representing correct
  answers of \code{HCItest} dataset.
}
\references{
McFarland, J. L., Price, R. M., Wenderoth, M. P., Martinkova, P.,
Cliff, W., Michael, J., ... & Wright, A. (2017). Development and validation
of the homeostasis concept inventory. CBE-Life Sciences Education, 16(2),
ar35. \doi{10.1187/cbe.16-10-0305}
}
\seealso{
\code{\link{HCItest}} for HCI multiple-choice dataset\cr
\code{\link{HCI}} for HCI dichotomous dataset\cr
\code{\link{HCIdata}} for HCI full dataset\cr
\code{\link{HCIgrads}}  for HCI dataset of graduate students\cr
\code{\link{HCIprepost}} for HCI pretest and posttest scores\cr
\code{\link{HCItestretest}} for HCI test-retest dataset
}
\author{
Jenny L. McFarland \cr Biology Department, Edmonds Community College
}
\keyword{datasets}
