% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_HCIprepost.R
\docType{data}
\name{HCIprepost}
\alias{HCIprepost}
\title{Homeostasis concept inventory pretest and posttest scores}
\format{
\code{HCIprepost} is a \code{data.frame} consisting of 16
observations on the 2 variables.
\describe{
\item{id}{Anonymized respondent ID. }
\item{score.pre}{Pretest score. } \item{score.post}{Posttest score. }
}
}
\usage{
HCIprepost
}
\description{
\code{HCIprepost} dataset consists of the pretest and
posttest score of 16 students to Homeostasis Concept Inventory (HCI).
Between the pre-test and post-test, the students received
instruction on homeostasis within a physiology course.
}
\references{
McFarland, J. L., Price, R. M., Wenderoth, M. P., Martinkova, P.,
Cliff, W., Michael, J., ... & Wright, A. (2017). Development and validation
of the homeostasis concept inventory. CBE-Life Sciences Education, 16(2),
ar35. \doi{10.1187/cbe.16-10-0305}
}
\seealso{
\link{HCI} for HCI dichotomous dataset\cr
\link{HCItest} for HCI multiple-choice dataset\cr
\link{HCIkey} for key of correct answers for HCI\cr
\link{HCIdata} for HCI full dataset\cr
\link{HCIlong} for HCI in a long format\cr
\link{HCIgrads}  for HCI dataset of graduate students\cr
\link{HCItestretest} for HCI test-retest dataset\cr
}
\author{
Jenny L. McFarland \cr Biology Department, Edmonds Community College
}
\keyword{datasets}
