% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ItemAnalysis.R
\name{ItemAnalysis}
\alias{ItemAnalysis}
\title{Item Analysis}
\usage{
ItemAnalysis(data, y = NULL, k = 3, l = 1, u = 3,
                    maxscore, minscore, cutscore, add.bin=FALSE)
}
\arguments{
\item{data}{matrix or data.frame of items to be examined. Rows represent respondents, columns
reperesent items.}

\item{y}{vector of criterion values.}

\item{k}{numeric: number of groups to which may be data.frame x divided by the total score.
Default value is 3.  See \strong{Details}.}

\item{l}{numeric: lower group. Default value is 1. See \strong{Details}.}

\item{u}{numeric: upper group. Default value is 3. See \strong{Details}.}

\item{maxscore}{numeric or vector: maximal score in ordinal items. If missing, vector of obtained maximal scores is imputed. See \strong{Details}.}

\item{minscore}{numeric or vector: minimal score in ordinal items. If missing, vector of obtained minimal scores is imputed. See \strong{Details}.}

\item{cutscore}{numeric or vector: cut score used for binarization of ordinal data.
If missing, vector of maximal scores is imputed. See \strong{Details}.}

\item{add.bin}{logical: If TRUE, indices are printed also for binarized data. See \strong{Details}.}
}
\description{
ItemAnalysis function computes various traditional item analysis indices
including difficulty, discrimination and item validity.
For ordinal items the difficulty and discrimination indices take into account
minimal item score as well as range.
}
\details{
ItemAnalysis function computes various traditional item analysis indices
including difficulty indices based on ratio of correct answers, sample SD,
discrimination indices ULI based on difference in ratio of correct answer
in Upper and Lower group, RIT (correlation between item score and overall test score),
RIR (correlation between item score and overall test score),
item validity indices based on correlation of Item score with criterion,
and item reliability indices based on Cronbach's alpha without given item.
For ordinal items the difficulty and discrimination indices take into account
minimal item score as well as range.

For caluclation discimination ULI index, it is possible to
specify number of groups \code{k}, and which two groups \code{l} and \code{u}
are to be compared.

In ordinal items, difficulty is calculated as difference of average score divided by range
(maximal possible score \code{maxscore} minus minimal possible score \code{minscore}).
If \code{add.bin} is set to \code{TRUE}, item analysis of binarized data is
included in the output table. In such a case, \code{cutscore} is used for binarization.
When binarizing the data, values greater or equal to cut-score are set to 1,
other values are set to 0.
}
\examples{
\dontrun{
## Examples on 100-item medical admission test data set (binary)
library(ShinyItemAnalysis)
dim(dataMedical)
names(dataMedical)

dataBin <- dataMedical[,1:100]
StudySuccessBin <- dataMedical[,102]

head(ItemAnalysis(dataBin))
head(ItemAnalysis(dataBin, y = StudySuccessBin))

## Examples on 100-item medical admission test data set (ordinal)
dim(dataMedicalgraded)
names(dataMedicalgraded)

dataOrd <- dataMedicalgraded[,1:100]
StudySuccess <- dataMedicalgraded[,102]

head(ItemAnalysis(dataOrd, y = StudySuccess))
head(ItemAnalysis(dataOrd, y = StudySuccess,
     k = 5, l = 4, u = 5, maxscore = 4, minscore = 0, cutscore = 4, add.bin=TRUE) )
}
}
\references{
Martinkova, P., Stepanek, L., Drabinova, A., Houdek, J., Vejrazka, M., & Stuka, C. (2017).
Semi-real-time analyses of item characteristics for medical school admission tests.
In: Proceedings of the 2017 Federated Conference on Computer Science and Information Systems.
http://dx.doi.org/10.15439/2017F380

Allen, M. J. & Yen, W. M. (1979). Introduction to measurement theory. Monterey, CA: Brooks/Cole.
}
\seealso{
\code{\link{DDplot}}, \code{\link{gDiscrim}}
}
\author{
Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr

Jana Vorlickova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr

Adela Drabinova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
drabinova@cs.cas.cz \cr
}
