% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blis.R
\name{print.blis_coefs}
\alias{print.blis_coefs}
\title{Print method for BLIS coefficients}
\usage{
\method{print}{blis_coefs}(x, digits = 3, ...)
}
\arguments{
\item{x}{result of \code{coef()}.}

\item{digits}{\emph{integer}, number of digits to show in the output. Note that
printed object are still an original list, which does not round any value
(it is returned invisibly).}

\item{...}{Additional arguments passed on to \code{print()}.}
}
\description{
Print method for BLIS coefficients
}
\seealso{
Other BLIS/BLIRT related: 
\code{\link{BlisClass-class}},
\code{\link{coef,BlisClass-method}},
\code{\link{fit_blis}()},
\code{\link{get_orig_levels}()},
\code{\link{nominal_to_int}()},
\code{\link{obtain_nrm_def}()}
}
\concept{BLIS/BLIRT related}
