% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startShinyItemAnalysis.R
\name{startShinyItemAnalysis}
\alias{startShinyItemAnalysis}
\title{This function will start ShinyItemAnalysis application.}
\usage{
startShinyItemAnalysis()
}
\description{
An interactive shiny application for running test and item analysis.
}
\examples{
\dontrun{
rm(list = ls())
startShinyItemAnalysis()
}
}
\author{
Patricia Martinkova \cr
Institute of Computer Science of the Czech Academy of Sciences \cr
\email{martinkova@cs.cas.cz} \cr

Adela Hladka \cr
Institute of Computer Science of the Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
\email{hladka@cs.cas.cz} \cr
}
