% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startShinyItemAnalysis.R
\name{startShinyItemAnalysis}
\alias{startShinyItemAnalysis}
\title{Start ShinyItemAnalysis application}
\usage{
startShinyItemAnalysis(background = TRUE)
}
\arguments{
\item{background}{logical: should the application be run as a background
process (in the 'RStudio')?}
}
\value{
No return value. Called for side effects.
}
\description{
An interactive shiny application to run test and item analysis.
By default, the function runs the application as a background process (Jobs
tab in the 'RStudio'). User is then free to use the Console for other work
and to try the sample R code examples. You can still run the app the usual
way in the console by specifying \code{background = FALSE}.
}
\examples{
\dontrun{
rm(list = ls())
startShinyItemAnalysis()
startShinyItemAnalysis(background = FALSE)
}

}
\author{
Patricia Martinkova \cr
Institute of Computer Science of the Czech Academy of Sciences \cr
\email{martinkova@cs.cas.cz} \cr

Adela Hladka \cr
Institute of Computer Science of the Czech Academy of Sciences \cr
\email{hladka@cs.cas.cz} \cr
}
