% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startShinyItemAnalysis.R
\name{startShinyItemAnalysis}
\alias{startShinyItemAnalysis}
\title{This function will start ShinyItemAnalysis application.}
\usage{
startShinyItemAnalysis()
}
\description{
An interactive shiny application for running test and item analysis.
}
\examples{
\dontrun{
rm(list = ls())
startShinyItemAnalysis()
}
}
\author{
Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr

Adela Drabinova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
adela.drabinova@gmail.com \cr

Ondrej Leder \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
ondraleder@gmail.com  \cr

Jakub Houdek \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
The Faculty of Informatics and Statistics, University of Economics, Prague \cr
houdek.james@gmail.com \cr
}
